requirements.txt
python-dotenv
langchain
langchain_openai
langchain_groq
langchain_chroma
langchain_community